........S. [ZBX-8542] removed local trigger cache
git-svn-id: svn://svn.zabbix.com/branches/dev@49288 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-8542/src/libs/zbxdbcache/dbcache.c
@@ -2187,7 +2187,7 @@ static void	DCvacuum_text()
 	zabbix_log(LOG_LEVEL_DEBUG, "In %s() text_free:%d/%d",
 			__function_name, cache->text_free, CONFIG_TEXT_CACHE_SIZE);
 
-	if (CONFIG_TEXT_CACHE_SIZE / 1024 >= cache->text_free)
+	if (CONFIG_TEXT_CACHE_SIZE / 1024 >= (unsigned int)cache->text_free)
 		goto exit;
 
 	cache->last_text = cache->text;
